Move KernelExp/KernelSigmoid to other_ops.rs and remove logical intermediaries
hlir.rs should only contain 1:1 HLIR op analogues. KernelExp and KernelSigmoid
are fused kernels, so they belong in other_ops.rs. Also removed the redundant
logical::Exp and logical::Sigmoid intermediary ops since the kernel ops match
HLIR patterns directly via their direct-fusion egglog rules.
